requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
bs4==0.0.1
captcha==0.2.4
cffi==1.11.5
cryptography==2.3.1
cycler==0.10.0
diff-match-patch==20121119
Django==1.11
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-ranged-response==0.2.0
django-reversion==3.0.0
django-simple-captcha==0.5.9
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
ipaddress==1.0.22
jdcal==1.4
jieba==0.39
kiwisolver==1.0.1
matplotlib==2.2.3
mysql-connector-python==8.0.13
numpy==1.15.3
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
